♻️ refactor: 优化 Episode Filter
